move builder to a new file
