+++ examples/builder-example/Cargo.toml
@@ -8,7 +8,6 @@ publish = false
 axum = "0.6.1"
 metrics = "0.20.1"
 metrics-exporter-prometheus = "0.11"
-pin-project-lite = "0.2.9"
 tokio = { version = "1.0", features = ["full"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
